Run lan server through babel-register so that pkg can understand it

+++ packages/lan/scripts/package.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 echo ${LAN_ROOT}
 
+if [ -z "$LAN_ROOT" ]; then
+  echo "Fatal error: LAN_ROOT not set."
+  exit 1
+fi
+
 rm -rf ${LAN_ROOT}/release && mkdir ${LAN_ROOT}/release
 yarn workspace lan run build
 cp ${LAN_ROOT}/.bin/*.node ${LAN_ROOT}/release/
